Added books for learning new moves.
Nick is working on some dummy art that will pop up to show you how to use your new move.
To specify which move you learn in Weltmeister, you specify "move": "doubleJump" (or whatever the move name is), which are based on the variable names of playerState in main.js.