Fixed sludge (see below)
The velocity was only being set once, on the frame where it went from "fromPuddle" to "walking".  It needs to be set every frame while walking in order for edge reverse to work.  Otherwise, it will just keep going at the same velocity that was set initially.  All edge reverse really does is toggle this.flip once an edge is reached.